Removed testify dependency from exported methods in testdata.go
